requirements.txt
tqdm==4.55.1
imageio==2.9.0
pandas==1.2.1
torch==1.5.1
matplotlib==3.1.3
numpy==1.19.2
torchvision==0.6.0a0+35d732a
seaborn==0.10.1
Pillow==8.1.0
termcolor==1.1.0
